Add support for multiple source sets
This feature can either be used to check additional source sets containing
production code, and/or it can be used to enable null checking on tests.

The source sets can be configured to be of type 'main' or of type 'test'.
If set to 'test', support for the AssertJ contract annotation is added,
and AssertJ calls influence the null checking.

Signed-off-by: Moritz Halbritter <moritz.halbritter@broadcom.com>